Make "Actions" accept floats for "x", "y" and "duration" properties (#303)
* Use floats as type for x, y and duration so that it doesn't break when a float is provided in Actions JSON
* When floats are provided in Actions JSON, round the values and provide warning to user if rounded value is different from provided value
* Also, add 'warn' as a log type (along with 'info', 'debug' and 'error')

+++ espresso-server/app/src/androidTest/java/io/appium/espressoserver/lib/handlers/exceptions/MoveTargetOutOfBoundsException.java
@@ -5,7 +5,7 @@ import io.appium.espressoserver.lib.helpers.Rect;
 
 public class MoveTargetOutOfBoundsException extends AppiumException {
 
-    public MoveTargetOutOfBoundsException(long targetX, long targetY, final Rect boundingRect) {
+    public MoveTargetOutOfBoundsException(float targetX, float targetY, final Rect boundingRect) {
         super(String.format(
             "The target [%s, %s] for pointer interaction is not in the viewport %s and cannot be brought into the viewport",
             targetX, targetY, boundingRect.toShortString()
